# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ChessCouncil",
    "description": "多 Agent 协作辩论的实时国际象棋分析与对战系统",
    "version": "0.4.1"
  },
  "paths": {
    "/api/game/new": {
      "post": {
        "summary": "New Game",
        "operationId": "new_game_api_game_new_post",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/game/move": {
      "post": {
        "summary": "Make Move",
        "operationId": "make_move_api_game_move_post",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/game/ai-step": {
      "post": {
        "summary": "Ai Step",
        "operationId": "ai_step_api_game_ai_step_post",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AiStepRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/game/state": {
      "get": {
        "summary": "Get State",
        "operationId": "get_state_api_game_state_get",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/game/load-fen": {
      "post": {
        "summary": "Load Fen",
        "operationId": "load_fen_api_game_load_fen_post",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/game/analyze-position": {
      "post": {
        "summary": "Analyze Position",
        "description": "分析当前局面（不走子），用于 Demo / 识谱后 Council。",
        "operationId": "analyze_position_api_game_analyze_position_post",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AnalyzePositionRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/game/post-review": {
      "post": {
        "summary": "Post Game Review",
        "description": "人人局等：终局后统一生成 Council 评价与复盘。",
        "operationId": "post_game_review_api_game_post_review_post",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/game/undo": {
      "post": {
        "summary": "Undo Move",
        "description": "悔棋（人 vs AI 尽量回到人类回合）。",
        "operationId": "undo_move_api_game_undo_post",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/game/hint": {
      "post": {
        "summary": "Hint Move",
        "description": "Stockfish 提示着法（不触发 Council）。",
        "operationId": "hint_move_api_game_hint_post",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/game/review": {
      "get": {
        "summary": "Game Review",
        "description": "赛后复盘报告（基于本局逐步 Council 缓存）。",
        "operationId": "game_review_api_game_review_get",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/game/save": {
      "post": {
        "summary": "Save Game",
        "operationId": "save_game_api_game_save_post",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/SaveGameRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fen/grid": {
      "get": {
        "summary": "Fen Grid",
        "operationId": "fen_grid_api_fen_grid_get",
        "parameters": [
          {
            "name": "fen",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fen"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fen/set-square": {
      "post": {
        "summary": "Fen Set Square",
        "operationId": "fen_set_square_api_fen_set_square_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FenSquareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AiStepRequest": {
        "properties": {
          "with_analysis": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "With Analysis"
          }
        },
        "type": "object",
        "title": "AiStepRequest"
      },
      "AnalyzePositionRequest": {
        "properties": {
          "with_analysis": {
            "type": "boolean",
            "title": "With Analysis",
            "default": true
          }
        },
        "type": "object",
        "title": "AnalyzePositionRequest"
      },
...
      "FenRequest": {
        "properties": {
          "fen": {
            "type": "string",
            "title": "Fen"
          }
        },
        "type": "object",
        "required": [
          "fen"
        ],
        "title": "FenRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MoveRequest": {
        "properties": {
          "uci": {
            "type": "string",
            "title": "Uci"
          },
          "with_analysis": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "With Analysis"
          },
          "analysis_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "fast",
                  "deep"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Mode"
          }
        },
        "type": "object",
        "required": [
          "uci"
        ],
        "title": "MoveRequest"
      },
...
    }
  }
}